Refactor formatting and tokenization utilities
