+++ shared/README.md
@@ -9,6 +9,7 @@ This folder contains utility modules shared across the major pipeline folders.
 | [`dataset_loader.py`](dataset_loader.py) | `DatasetLoader` | Loads datasets from local files, directories, or HuggingFace Hub. |
 | [`logging.py`](logging.py)               | `get_logger`    | Creates a logger with a consistent format.                        |
 | [`save_dataset.py`](save_dataset.py)     | `save_dataset`  | Saves a dataset to disk, splitting into token-counted chunks.     |
+| [`files.py`](files.py) | `list_matching_files`, `infer_file_features` | Glob file listing and feature inference from parquet/jsonl shards. |
 
 ## Usage
