+++ shared/__init__.py
@@ -4,4 +4,5 @@ Modules:
     - dataset_loader: DatasetLoader
     - logging: get_logger
     - save_dataset: save_dataset
+    - files: list_matching_files, infer_file_features
 """
